separate paging and pmem
