+++ src/include/kernel/addr.h
@@ -1,0 +1,8 @@
+#ifndef __ARGIR__ADDR_H
+#define __ARGIR__ADDR_H
+
+#define KERNEL_LMA (0x200000)
+#define KERNEL_VMA (0xffffffff80000000)
+#define LFB_VMA (0xffffffff40200000) /** Only available after physmem init */
+
+#endif /** __ARGIR__ADDR_H */
