+++ src/kernel.c
@@ -2,6 +2,7 @@
 #include <stddef.h>
 #include <stdbool.h>
 #include <stdio.h>
+#include "kernel/addr.h"
 #include "kernel/mb2.h"
 #include "kernel/cpu.h"
 #include "kernel/interrupts.h"
@@ -14,11 +15,44 @@
 #define __ARGIR_BUILD_COMMIT__ "balls"
 #endif
 
-struct pci g_pci;
 uint32_t mb2_magic;
 uint32_t mb2_info;
 
-static void print_logo()
+static void print_build_info();
+
+void kernel_main(void)
+{
+    interrupts_disable();
+    if (mb2_magic != 0x36d76289) {
+        __asm__ volatile("mov $0xbaaaaaadb0000007, %rax\n\t"
+                         "hlt"); // MB2 did not drop us here!
+    }
+
+    // Extract required info from MB2 bootinfo before it disappears
+    struct mb2_tag *mb2_tag_fb = mb2_find_tag(mb2_info, 8);
+    size_t width = mb2_tag_fb->framebuffer.width;
+    size_t height = mb2_tag_fb->framebuffer.height;
+    size_t pitch = mb2_tag_fb->framebuffer.pitch;
+
+    pmem_init(mb2_info);
+    terminal_init(LFB_VMA, width, height, pitch);
+    print_build_info();
+
+    gdt_init();
+    interrupts_init();
+    keyboard_init();
+
+    // Ready to go
+    interrupts_enable();
+
+    for (;;) {
+        keyboard_main();
+
+        __asm__ volatile("hlt");
+    }
+}
+
+static void print_build_info()
 {
     printf(
         "\n                                @@\\\n"
@@ -34,39 +68,4 @@ static void print_logo()
         "                       \\______/\n\n"
         "Argir x86_64\n"
         "Build " __ARGIR_BUILD_COMMIT__ "\n\n");
-}
-
-void kernel_main()
-{
-    // Sanity check!
-    interrupts_disable();
-    if (mb2_magic != 0x36d76289) {
-        /** wat do?! */
-        __asm__ volatile("mov $0xdeadbeef, %rax\n\t"
-                         "hlt");
-    }
-
-    // Graphics
-    struct mb2_tag *mb2_tag_fb = mb2_find_tag(mb2_info, 8);
-    size_t width = mb2_tag_fb->framebuffer.width;
-    size_t height = mb2_tag_fb->framebuffer.height;
-    size_t pitch = mb2_tag_fb->framebuffer.pitch;
-    pmem_init(mb2_info);
-    uint8_t *fb = 0xffffffff40200000;
-    terminal_init(fb, width, height, pitch);
-    print_logo();
-
-    // Early init
-    gdt_init();
-    interrupts_init();
-    keyboard_init();
-
-    // Ready to go
-    interrupts_enable();
-
-    for (;;) {
-        keyboard_main();
-
-        __asm__ volatile("hlt");
-    }
 }
